Use 'use_inline_resources' to ensure notifications propagate correctly
PR #69 updated the monit provider to use run_action on the provided
resources to make external notifications propagate (i.e., so that a
'notifies' in a monit_monitrc's block would run). But that has the side
effect of not firing the 'notifies' inside the monit_monitrc resource,
and thus monit never reloads when a new monit_monitrc resource is
created.

This use use_inline_resources to evaluate the monit template in its own
context, so that that context will fire notifications declared on the
monit_monitrc resource, while still firing the notification on the
template itself.

See notes on updated_by_last_action here:
  https://docs.chef.io/custom_resources_notes.html